refactor: move mockLogger into the specific test file
