+++ packages/better-auth/src/api/check-endpoint-conflicts.test.ts
@@ -1,11 +1,32 @@
-import { describe, it, expect, vi } from "vitest";
+import { describe, it, expect, vi, beforeEach } from "vitest";
 import { checkEndpointConflicts } from "./index";
 import type { BetterAuthOptions, BetterAuthPlugin } from "../types";
 import { createEndpoint } from "better-call";
-import { mockLogger } from "../test/utils";
+import type { InternalLogger, LogLevel } from "../utils";
+
+export let mockLoggerLevel: LogLevel = "debug";
+export const mockLogger = {
+	error: vi.fn(),
+	warn: vi.fn(),
+	info: vi.fn(),
+	debug: vi.fn(),
+	success: vi.fn(),
+	get level(): LogLevel {
+		return mockLoggerLevel;
+	},
+} satisfies InternalLogger;
 
 describe("checkEndpointConflicts", () => {
 	const endpoint = createEndpoint.create({});
+
+	beforeEach(() => {
+		mockLoggerLevel = "debug";
+		mockLogger.error.mockReset();
+		mockLogger.warn.mockReset();
+		mockLogger.info.mockReset();
+		mockLogger.debug.mockReset();
+		mockLogger.success.mockReset();
+	});
 
 	it("should not log errors when there are no endpoint conflicts", () => {
 		const plugin1: BetterAuthPlugin = {
